Add hook for speed limit zones' data
